+++ tests/test_lazy.py
@@ -1,6 +1,6 @@
 from textual.app import App, ComposeResult
 from textual.containers import Horizontal, Vertical
-from textual.lazy import Lazy
+from textual.lazy import Lazy, Reveal
 from textual.widgets import Label
 
 
@@ -24,3 +24,32 @@ async def test_lazy():
         # #bar mounted after refresh
         assert len(app.query("#foo")) == 1
         assert len(app.query("#bar")) == 1
+
+
+class RevealApp(App):
+    def compose(self) -> ComposeResult:
+        with Reveal(Vertical()):
+            yield Label(id="foo")
+            yield Label(id="bar")
+            yield Label(id="baz")
+
+
+async def test_lazy_reveal():
+    app = RevealApp()
+    async with app.run_test() as pilot:
+        # No #foo on initial mount
+
+        # Only first child should be visible initially
+        assert app.query_one("#foo").display
+        assert not app.query_one("#bar").display
+        assert not app.query_one("#baz").display
+
+        # All children should be visible after a pause
+        await pilot.pause()
+        for n in range(3):
+            await pilot.pause(1 / 60)
+            await pilot.pause()
+
+        assert app.query_one("#foo").display
+        assert app.query_one("#bar").display
+        assert app.query_one("#baz").display