+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fixed duplicated key displays in the help panel https://github.com/Textualize/textual/issues/5037
+- Fixed crash when removing then adding panes to `TabbedContent` https://github.com/Textualize/textual/issues/5215
 
 ## [0.85.2] - 2024-11-02